Add files via upload
https://sites.google.com/site/brucehuntsgroundwaterwebsite/
Bruce Hunt's Groundwater Website:

Function is an Excel spreadsheet program that contains a large collection of user-defined programs for the solution of some specific problems in groundwater resources. These programs are implemented by entering in an Excel spreadsheet cell the call name and list of variables used in the solution of each particular problem. (Please note that these are not fill-in-the-blanks spreadsheets. Instead, users must use their own ingenuity to design spreadsheet programs for their own particular problems.)
Groundwater Analysis Using Function contains physical descriptions, mathematical descriptions, computer call statements and some worked examples for groundwater programs contained in Function.
FUNCTION IN NEWER VERSIONS OF EXCEL explains how to enable the macros in Function.xls to be used in Excel 2007 and Excel 2010.
Review of Stream Depletion Solutions, Behavior, and Calculations is a paper that was published in the ASCE Journal of Hydrologic Engineering in January 2014.